requirements.txt
websocket-client==0.47.0
